Safe node: do not put object values in fn-validation ex-data
In the worst case, putting the invalid value in the exception data can
cause stack overflow errors which are very
costly and troublesome to debug if/when they ever do happen. For
example, this can happen by calling a validated function on any
structure that contains a self-referencing object which tries to print.